requirements.txt
# ==== 런타임 필수 ====
fastapi~=0.110
uvicorn[standard]~=0.29
httpx[http2]~=0.27
beautifulsoup4~=4.12
pydantic~=2.7
supabase~=2.4        # Supabase Python 클라이언트
lxml~=5.2            # BeautifulSoup 'lxml' 파서

# ==== 크롤러 & LLM ====
playwright~=1.40     # Naver Map 크롤링
google-generativeai~=0.3  # Gemini LLM (선택)
python-dotenv~=1.0   # 환경변수 로드

# ==== 테스트 ====
pytest~=8.2
pytest-asyncio~=0.23

# ==== 선택: 실제 코드에서 사용 중이면 주석 해제 ====
# prometheus-client~=0.20    # monitor/metrics 사용시
# sentry-sdk~=2.0            # Sentry 연동시
# redis~=5.0                 # redis 캐시/레이트리밋 사용시
# fastapi-limiter~=0.1       # redis 기반 레이트리밋 사용시
lxml~=5.2                  # BeautifulSoup에서 'lxml' 파서 지정 시

slowapi~=0.1.9
